GnuPG plugin for Vim
This script implements transparent editing of gpg encrypted files. The
filename must have a ".gpg" suffix. When opening such a file the content
is decrypted, when opening a new file the script will ask for the
recipients of the encrypted file. The file content will be encrypted to
all recipients before it is written. The script turns off viminfo and
swapfile to increase security.